+++ data/src/main/kotlin/team/aliens/dms/android/data/latestudy/di/LateStudyRepositoryModule.kt
@@ -5,7 +5,7 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import team.aliens.dms.android.data.latestudy.repository.LateStudyRepositoryImpl
-import team.aliens.dms.android.domain.latestudy.repository.LateStudyRepository
+import team.aliens.dms.android.data.latestudy.repository.LateStudyRepository
 
 @Module
 @InstallIn(SingletonComponent::class)
