+++ data/src/main/kotlin/team/aliens/dms/android/data/latestudy/repository/LateStudyRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package team.aliens.dms.android.data.latestudy.repository
 
 import team.aliens.dms.android.data.latestudy.mapper.toModel
-import team.aliens.dms.android.domain.latestudy.model.StudyType
-import team.aliens.dms.android.domain.latestudy.repository.LateStudyRepository
+import team.aliens.dms.android.data.latestudy.model.StudyType
+import team.aliens.dms.android.data.latestudy.repository.LateStudyRepository
 import team.aliens.dms.android.network.latestudy.datasource.NetworkLateStudyDataSource
 import javax.inject.Inject
