+++ network/src/main/kotlin/team/aliens/dms/android/network/latestudy/datasource/NetworkLateStudyDataSourceImpl.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.android.network.latestudy.datasource
 
 import team.aliens.dms.android.network.latestudy.apiservice.LateStudyApiService
 import team.aliens.dms.android.network.latestudy.model.FetchStudyTypesResponse
-import team.aliens.dms.android.network.latestudy.model.response.FetchStudyTypesResponse
+
 import javax.inject.Inject
 
 class NetworkLateStudyDataSourceImpl @Inject constructor(
